Move bisect-perf-regression.py into auto_bisect directory.
The purpose of this is to continue gathering all the related code together, and to make it easier to run pylint/tests for the main module which still contains most of the code.

After doing this, pylint and the tests caught a couple of things -- one unused variable name, and after the script changed directory, some of the paths had to be modified as well.

BUG=

Review URL: https://codereview.chromium.org/564663002

Cr-Original-Commit-Position: refs/heads/master@{#296847}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62edbcf1e8e5dd1386693b7b4b90169e72bd7300