[Telemetry] Properly chose an available device port.
Previously, we'd start WPR on an available host port and then just
assume that the same port was free on the device -- oops.

Now, we pass 0 to the forwarder for the device port so that it chooses
an available port, and then read back the port that it chose to use
for the mapping.

I believe this fixes a major case of device forwarder flake.

BUG=421599
TBR=slamm@chromium.org

Review URL: https://codereview.chromium.org/640133002

Cr-Original-Commit-Position: refs/heads/master@{#299319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6aa01dbdfbce3a03522f8bc0b59916529de54a93