[Telemetry/Mac] Don't crash when setuid bit not set for powermetrics
When Telemetry is run it will prompt to set the setuid bit for powermetrics, if the user opts not to do this, Telemetry will crash because the Python call to terminate() doesn't have appropriate privileges to kill the process.

The fix is to send SIGTERM via `sudo kill`.

BUG=None

Review URL: https://codereview.chromium.org/654803002

Cr-Original-Commit-Position: refs/heads/master@{#299506}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: abb19226cb71da65d798ef3d18527331bc349605