Add getUserSelectedDevices to the USB extensions API.
chrome.usb.getUserSelectedDevices allows an app to prompt the user for
one or more USB devices for it to access. This is an alternative to
statically declaring the list of devices an app can access in the
application manifest. This is useful for device classes such as
ADB-enabled Android devices which may have a great number of vendor and
product ID pairs but all match a well-known interface protocol.

If a selected device has a serial number then the app can retain access
to it until explicitly revoked (as is the case for most Android phones)
while access to a device without a serial number is revoked when the app
is unloaded or the device is disconnected.

BUG=346953

Review URL: https://codereview.chromium.org/599303004

Cr-Original-Commit-Position: refs/heads/master@{#300141}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 042073be6f85951049a6a3a4f84499ab85109566