Look for gclient.py in find_depot_tools
Many (but not all!) of our bots have a directory early in the PATH that
contains nothing but a script called 'gclient' and have the real
depot_tools directory later on in the path. This looks for gclient.py,
which is in the real depot_tools directory.

Review URL: https://codereview.chromium.org/652893005

Cr-Original-Commit-Position: refs/heads/master@{#300802}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 380a535523ed8b808eab6fa7c205dce67e0ed2a9