Wait 60s instead of 30s for browser to launch.
There's a guess that the browser may be launching extremely slowly on
some bots and that when the browser main process is killed it's
leaking sub-processes that are in the middle of being spawned.
Increase the timeout to try to confirm this hypothesis.

BUG=424024
R=tonyg@chromium.org

Review URL: https://codereview.chromium.org/691683002

Cr-Original-Commit-Position: refs/heads/master@{#302087}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ef2c0036f70bdcc59dd7c8f23c8b7c5a3f114d9