Collect license version as an optional_key in licenses.py
This changes licenses.py to collect the license version information
during ParseDir, as an optional field.  This is for scripts that
import this script as a module, that may need this information.

Additionally, this removes the "Required Text" support from
licenses.py, because it was only used by SwiftShader, which no
longer needs this functionality.

BUG=424424

Review URL: https://codereview.chromium.org/660163002

Cr-Original-Commit-Position: refs/heads/master@{#302213}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95468c6efabd1ec13b9926c762f2021400a14192