Add minimum TLS version control to about:flags and Finch gate it.
This change adds an entry to about:flags to control the minimum TLS version as
requested by Darin. This will allow Android and ChromeOS users to reenable
SSLv3 in Chrome 40 if they need to.

Additionally, this gates the SSLv3 disabling with Finch. At the moment the
experiment doesn't actually exist and so will always return the empty string.
However, should we need to, we can define it and put everyone in the Enabled
group.

BUG=427671,419870

Review URL: https://codereview.chromium.org/693963003

Cr-Original-Commit-Position: refs/heads/master@{#302754}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a4a16346d0a921949ad8f423934350a3b504e70