[Telemetry] Fix accidentally enabled benchmarks.
@Enabled is satisfied if ANY of the fields match. The intent behind these
particular uses was to be enabled if ALL of the fields match. So for now, I'm
disabling on desktop rather than enabling on Android. But in the future, we
should probably have a better way of indicating ANY vs ALL.

BUG=426718

Review URL: https://codereview.chromium.org/733263002

Cr-Original-Commit-Position: refs/heads/master@{#304537}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ae48cc1358ff6367a0645f6027f3303266c21d7