cc: Add UMA stats for texture copy latency.
Records the time between the issue of a texture copy operation and
detecting that it has completed as measured by the GPU process.

This data is useful to determine potential driver problems related
to fences. Consistent high latency (200ms and above) for copy
operations is a strong indication of a misbehaving fence extension.

The data can also be used to determine the amount of staging
resources to use for optimal performance.

BUG=431845

Review URL: https://codereview.chromium.org/733053005

Cr-Original-Commit-Position: refs/heads/master@{#304562}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 63294bc251336f3e59240642676a5e758a8c2192