Measure video stats for WebRTC calls to ensure video works.
This will count the number of dropped frames and the number of decoded
frames. This provides a reasonable guarantee that the video pipeline is
working.

The next step would be to look at the actual pixels in the video.

BUG=434701

Review URL: https://codereview.chromium.org/738973005

Cr-Original-Commit-Position: refs/heads/master@{#305030}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56c60209b09297951f2321d5b5818e1a97dca465