De-dup PrefixSet code in SafeBrowsingDatabaseManager.
A lot of SideEffectWhiteListPrefixSet related code was mostly a duplicate
 of the other PrefixSet code in the same file.

Only logical side-effect this has is grouping histograms for all PrefixSets
in the same bucket. This will be addressed as a follow-up in
https://codereview.chromium.org/815863002/ where relevant histograms will be
split once again.

This cleanup makes it easier to progress on issue 440517.

BUG=440517
TBR=jwd

Review URL: https://codereview.chromium.org/790703003

Cr-Original-Commit-Position: refs/heads/master@{#309589}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe11d94a8cb3e3b7d942d1cb3f654c75fe664678