Expect resources to have DRP via header in header_validation test.
Change the DRP header_validation telemetry test to expect that the DRP via header is present on all eligible responses. This makes it more consistent with the other tests, and causes it not to crash on Android Webview when the DRP is disabled.

BUG=446343

Review URL: https://codereview.chromium.org/829393003

Cr-Original-Commit-Position: refs/heads/master@{#310213}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b94a8563e3fe462e1bafb7e0f183f514a6943227