Move chrome_browser_backend's StopTracing logic to DevToolsClientBackend fully.
This paves the way to modify tracing_controller_backend to not rely on
browser_backend (and instead rely on DevToolsClientBackend) to issue
chrome tracing request.

This is accompanied with a rather major refactoring of
InspectorBackendList. Now the responsibility of creating and caching
InspectorBackendList belong solely to
DevToolsClientBackend/_DevToolsContextMapBackend.

InspectorBackendList is now in charge of filtering and creating wrapper.

BUG=423954,443897

Review URL: https://codereview.chromium.org/810883004

Cr-Original-Commit-Position: refs/heads/master@{#310860}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5d4da02e41b4c8de9ee03d818eee94e2b8888ef2