Remove dependency on dateutil in memory_inspector
This patch is part of an ongoing process to unbundle memory_inspector
and make it a chrome app. It removes the dependency of www_server.py
on the dateutil module, which is not part of the Python Standard
Library.

BUG=448399

Review URL: https://codereview.chromium.org/850033002

Cr-Original-Commit-Position: refs/heads/master@{#311507}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2ddff339b83714e75c8da7d0be99dbc3b4c3213b