Restart browser between smoke tests, fixes failing webview bypass test.
The bypass smoke test uses an existing browser session, where the favicon requewst has already been made and completed through the data saver proxy.  When the favicon is requested again during the bypass test, it will still appear to come through the proxy causing the case to fail.  If instead we use a fresh browser session, bypass works as expected for the favicon.

In general, it's good practice to make tests as atomic as possible, so this should be better design for the smoke tests (although it adds ~12 seconds to the total execution time).

BUG=452161

Review URL: https://codereview.chromium.org/862803004

Cr-Original-Commit-Position: refs/heads/master@{#313153}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 52dc983774bcd3262472f025d0429e10799c3f80